+++ frontend/src/components/EditCV.jsx
@@ -1,239 +1,225 @@
-/*
- This component serve the purpose to provide an interface to render the CV fields retrieved from the back, it consumes an static JSON object to know which fields to render and their values, it also uses a translation object to render the labels in the correct language, and a placeholder object to provide examples of the expected input for each field.
-*/
 import React, { useState, useEffect } from "react";
 import RAW_JSON from "@data/mock/RawJson";
 import TRANSLATIONS from "@data/CV/Translations";
 import CV_PLACEHOLDERS from "../data/CV/Placeholders";
 
-// Sections allowed to have dynamic add/remove functionality (Mapped from RAW_JSON)
-const ALLOWED_SECTIONS = [
-  "profiles",
-  "work",
-  "volunteer",
-  "education",
-  "awards",
-  "certificates",
-  "publications",
-  "projects",
-];
+const SECTIONS = Object.keys(RAW_JSON);
 
 function EditCV() {
-  const [data, setData] = useState(null);
+    const [data, setData] = useState({});
+    const [step, setStep] = useState(0);
 
-  // Simulate fetching data with a delay
-  useEffect(() => {
-    setTimeout(() => {
-      setData(RAW_JSON);
-    }, 1000);
-  }, []);
+    useEffect(() => {
+        setData(JSON.parse(JSON.stringify(RAW_JSON)));
+    }, []);
 
-  // Handles changes to input fields and updates state
-  const handleChange = (path, value) => {
-    setData((prevData) => {
-      const newData = JSON.parse(JSON.stringify(prevData));
-      let ref = newData;
-      const keys = path.split(".");
-      keys.slice(0, -1).forEach((key) => {
-        if (!ref[key]) ref[key] = {};
-        ref = ref[key];
-      });
-      ref[keys[keys.length - 1]] = value;
-      return newData;
-    });
-  };
-
-  const handleAddItem = (path) => {
-      setData((prevData) => {
-        const newData = JSON.parse(JSON.stringify(prevData)); // Deep copy of current state
-        const baseTemplate = JSON.parse(JSON.stringify(RAW_JSON)); // Deep copy of RAW_JSON
-        let ref = newData;
-        let templateRef = baseTemplate;
-  
-        // Traverse the path to find the correct section
-        const keys = path.split(".");
-        keys.forEach((key, index) => {
-          if (!ref[key]) ref[key] = index === keys.length - 1 ? [] : {};
-          ref = ref[key];
-          templateRef = templateRef[key]; // Get the corresponding base structure
+    const handleChange = (section, index, key, value) => {
+        setData((prevData) => {
+            const newData = JSON.parse(JSON.stringify(prevData));
+            if (Array.isArray(newData[section])) {
+                newData[section][index][key] = value;
+            } else {
+                newData[section][key] = value;
+            }
+            return newData;
         });
-  
-        // Push a new item based on the base structure
-        if (Array.isArray(templateRef)) {
-          ref.push(JSON.parse(JSON.stringify(templateRef[0]))); // Clone the first item in the base structure
-        } else {
-          console.error("Template for the section is not an array.");
-        }
-  
-        return newData;
-      })
-  };
+    };
+
+    const handleAddItem = (section) => {
+        setData((prevData) => {
+            const newData = JSON.parse(JSON.stringify(prevData));
+            if (!Array.isArray(newData[section])) newData[section] = [];
+            const template = Array.isArray(RAW_JSON[section]) ? RAW_JSON[section][0] || {} : {};
+            newData[section].push(JSON.parse(JSON.stringify(template)));
+            return newData;
+        });
+    };
+
+    const handleRemoveItem = (section, index) => {
+        setData((prevData) => {
+            const newData = JSON.parse(JSON.stringify(prevData));
+            if (Array.isArray(newData[section])) {
+                newData[section].splice(index, 1);
+            }
+            return newData;
+        });
+    };
+
+    const handleSubmit = () => {
+        console.log("Submitted Data:", JSON.stringify(data, null, 2));
+    };
 
 
-  // Removes an item from an array-based section
-  const handleRemoveItem = (path, index, event) => {
-    event.preventDefault();
-    setData((prevData) => {
-      const newData = JSON.parse(JSON.stringify(prevData));
-      let ref = newData;
-      const keys = path.split(".");
-      keys.forEach((key) => {
-        ref = ref[key];
-      });
-      ref.splice(index, 1); // Remove item at specified index
-      return newData;
-    });
-  };
+    if (!data) return <div>Loading...</div>;
 
-  // Handles form submission
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    console.log(data);
-  };
-
-  if (!data) return <div>Loading...</div>;
-
-  return (
-    <div className="container-fluid w-100 vh-100 p-4 bg-light">
-      <h1 className="text-center mb-4">Datos extraídos</h1>
-      <form>
-        {renderFields(data, "", handleChange, handleAddItem, handleRemoveItem)}
-        <div className="w-100 d-flex justify-content-center pb-3">
-          <button
-            type="submit"
-            className="btn btn-primary w-50"
-            onClick={handleSubmit}
-          >
-            Continuar &gt;
-          </button>
-        </div>
-      </form>
-    </div>
-  );
-}
-
-// Recursively renders form fields
-function renderFields(
-  data,
-  path,
-  handleChange,
-  handleAddItem,
-  handleRemoveItem,
-) {
-  return Object.keys(data).map((key) => {
-    const value = data[key];
-    const newPath = path ? `${path}.${key}` : key;
-    const label = TRANSLATIONS[key] || key;
-    const placeholder = getPlaceholder(newPath);
-
-    //If it's an object call renderFields recursively
-    if (typeof value === "object" && !Array.isArray(value)) {
-      return (
-        <fieldset
-          key={newPath}
-          className="border p-3 mb-3 bg-white shadow-sm rounded"
-        >
-          <legend className="fw-bold text-uppercase">
-            <h3>{label}</h3>
-          </legend>
-          {renderFields(
-            value,
-            newPath,
-            handleChange,
-            handleAddItem,
-            handleRemoveItem,
-          )}
-        </fieldset>
-      );
-      // If its an array, render the title and call renderFields for each item
-    } else if (Array.isArray(value) && ALLOWED_SECTIONS.includes(key)) {
-      const template =
-        value.length > 0 ? JSON.parse(JSON.stringify(value[0])) : {};
-      return (
-        <div key={newPath} className="mb-3">
-          <label className="fw-bold text-capitalize">
-            <h3>{label}</h3>
-          </label>
-          {value.map((item, index) =>
-            typeof item === "object" ? (
-              <fieldset
-                key={`${newPath}[${index}]`}
-                className="border p-3 bg-white shadow-sm rounded"
-              >
-                <legend className="fw-bold text-uppercase">{`${label} #${index + 1}`}</legend>
-                {renderFields(
-                  item,
-                  `${newPath}[${index}]`,
-                  handleChange,
-                  handleAddItem,
-                  handleRemoveItem,
+    return (
+        <div className="container-fluid w-100 vh-100 p-4 bg-light">
+            <h1 className="text-center mb-4 mt-4">Editar CV</h1>
+            <div className="d-flex justify-content-center mb-4">
+                {SECTIONS.map((_, index) => (
+                    <div
+                        key={index}
+                        className={`mx-1 w-3 h-3 rounded-full ${
+                            step === index ? "bg-secondary scale-125" : "bg-secondary"
+                        }`}
+                    ></div>
+                ))}
+            </div>
+            <form>
+                {SECTIONS[step] && data[SECTIONS[step]] && (
+                    <fieldset className="p-3 bg-white shadow-sm rounded">
+                        <legend className="fw-bold">
+                            {TRANSLATIONS[SECTIONS[step]] || SECTIONS[step]}
+                        </legend>
+                        {renderSection(
+                            data[SECTIONS[step]],
+                            SECTIONS[step],
+                            handleChange,
+                            handleAddItem,
+                            handleRemoveItem
+                        )}
+                    </fieldset>
                 )}
-                <button
-                  className="btn btn-danger mt-2"
-                  onClick={(e) => handleRemoveItem(newPath, index, e)}
-                >
-                  -
-                </button>
-              </fieldset>
-            ) : (
-              <input
-                key={`${newPath}[${index}]`}
-                type="text"
-                className="form-control mb-2 bg-secondary"
-                placeholder={`Ej: ${placeholder}`}
-                value={item || ""}
-                onChange={(e) => {
-                  const newValue = [...value];
-                  newValue[index] = e.target.value;
-                  handleChange(newPath, newValue);
-                }}
-              />
-            ),
-          )}
-          <button
-           type="button"
-            className="btn btn-success mt-2"
-            onClick={(e) => handleAddItem(newPath, template, e)}
-          >
-            +
-          </button>
+                <div className="w-100 d-flex justify-content-between mt-4">
+                    <button
+                        type="button"
+                        className="btn btn-outline-secondary"
+                        disabled={step === 0}
+                        onClick={() => setStep((prev) => prev - 1)}
+                    >
+                        &lt; Anterior
+                    </button>
+                    <button
+                        type="button"
+                        className="btn btn-primary"
+                        onClick={() => {
+                            if (step < SECTIONS.length - 1) {
+                                setStep((prev) => prev + 1);
+                            } else {
+                                handleSubmit();
+                            }
+                        }}
+                    >
+                        {step === SECTIONS.length - 1 ? "Finalizar" : "Siguiente >"}
+                    </button>
+                </div>
+            </form>
         </div>
-      );
-      // If its a key-value pair, render the label and an input field
-    } else {
-      return (
-        <div key={newPath} className="mb-1">
-          <label className="fw-bold text-capitalize">
-            <h3>{label}</h3>
-          </label>
-          <input
-            type="text"
-            className="form-control bg-secondary"
-            value={value || ""}
-            onChange={(e) => handleChange(newPath, e.target.value)}
-            placeholder={`Ej: ${placeholder}`}
-          />
-        </div>
-      );
-    }
-  });
+    );
 }
 
-// Retrieves placeholder text for inputs
-function getPlaceholder(path) {
-  const keys = path.split(".");
-  let ref = CV_PLACEHOLDERS;
+function renderSection(sectionData, section, handleChange, handleAddItem, handleRemoveItem) {
+    if (!sectionData) return null;
 
-  for (const key of keys) {
-    if (Array.isArray(ref)) {
-      ref = ref[0] || {};
-    } else if (typeof ref === "object" && ref[key]) {
-      ref = ref[key];
-    } else {
-      return "No disponible";
+    if (Array.isArray(sectionData)) {
+        return (
+            <div className="mb-3">
+                {sectionData.map((item, index) => (
+                    <div key={index} className="mb-2 p-2 border rounded">
+                        {Object.keys(item).map((key) =>
+                            typeof item[key] === "object" && key !== "profiles" ? (
+                                // Render nested objects (e.g., complex fields)
+                                <div key={key} className="mb-3">
+                                    <label className="fw-bold">{TRANSLATIONS[key] || key}</label>
+                                    {Object.keys(item[key]).map((subKey) => (
+                                        <input
+                                            key={subKey}
+                                            type="text"
+                                            className="form-control mb-2"
+                                            value={item[key][subKey] || ""}
+                                            onChange={(e) => handleChange(section, index, key, { ...item[key], [subKey]: e.target.value })}
+                                            placeholder={TRANSLATIONS[subKey] || subKey}
+                                        />
+                                    ))}
+                                </div>
+                            ) : (
+                                // Render simple fields
+                                <input
+                                    key={key}
+                                    type="text"
+                                    className="form-control mb-2"
+                                    value={item[key] || ""}
+                                    onChange={(e) => handleChange(section, index, key, e.target.value)}
+                                    placeholder={TRANSLATIONS[key] || key}
+                                />
+                            )
+                        )}
+                        <button type="button" className="btn btn-danger" onClick={() => handleRemoveItem(section, index)}>
+                            -
+                        </button>
+                    </div>
+                ))}
+                <button type="button" className="btn btn-success mt-2" onClick={() => handleAddItem(section)}>
+                    +
+                </button>
+            </div>
+        );
     }
-  }
-  return typeof ref === "string" ? ref : "No disponible";
+
+    return Object.keys(sectionData).map((key) =>
+        Array.isArray(sectionData[key]) && key === "profiles" ? (
+            // Render profiles (special case for "profiles" array)
+            <div key={key} className="mb-3">
+                <label className="fw-bold">{TRANSLATIONS[key] || key}</label>
+                {sectionData[key].map((profile, index) => (
+                    <div key={index} className="mb-2 p-2 border rounded">
+                        {Object.keys(profile).map((subKey) => (
+                            <input
+                                key={subKey}
+                                type="text"
+                                className="form-control mb-2"
+                                value={profile[subKey] || ""}
+                                onChange={(e) =>
+                                    handleChange(section, index, key, { ...profile, [subKey]: e.target.value })
+                                }
+                                placeholder={TRANSLATIONS[subKey] || subKey}
+                            />
+                        ))}
+                        <button
+                            type="button"
+                            className="btn btn-danger"
+                            onClick={() => handleRemoveItem(key, index)}
+                        >
+                            -
+                        </button>
+                    </div>
+                ))}
+                <button type="button" className="btn btn-success mt-2" onClick={() => handleAddItem(key)}>
+                    +
+                </button>
+            </div>
+        ) : key === "location" && typeof sectionData[key] === "object" ? (
+            // Render location (special case for "location" object)
+            <div key={key} className="mb-3">
+                <label className="fw-bold">{TRANSLATIONS[key] || key}</label>
+                {Object.keys(sectionData[key]).map((subKey) => (
+                    <input
+                        key={subKey}
+                        type="text"
+                        className="form-control mb-2"
+                        value={sectionData[key][subKey] || ""}
+                        onChange={(e) =>
+                            handleChange(section, 0, key, { ...sectionData[key], [subKey]: e.target.value })
+                        }
+                        placeholder={TRANSLATIONS[subKey] || subKey}
+                    />
+                ))}
+            </div>
+        ) : (
+            // Render other normal fields
+            <div key={key} className="mb-3">
+                <label className="fw-bold">{TRANSLATIONS[key] || key}</label>
+                <input
+                    type="text"
+                    className="form-control"
+                    value={sectionData[key] || ""}
+                    onChange={(e) => handleChange(section, 0, key, e.target.value)}
+                    placeholder={TRANSLATIONS[key] || key}
+                />
+            </div>
+        )
+    );
+    
 }
 
 export default EditCV;
